+++ internal/services/compute/orchestrated_virtual_machine_scale_set_resource_test.go
@@ -1,96 +1,12 @@
 package compute_test
 
 import (
-	"context"
 	"fmt"
 	"testing"
 
-	"github.com/Azure/azure-sdk-for-go/services/compute/mgmt/2021-07-01/compute"
-	"github.com/hashicorp/terraform-provider-azurerm/helpers/azure"
 	"github.com/hashicorp/terraform-provider-azurerm/internal/acceptance"
 	"github.com/hashicorp/terraform-provider-azurerm/internal/acceptance/check"
-	"github.com/hashicorp/terraform-provider-azurerm/internal/clients"
-	"github.com/hashicorp/terraform-provider-azurerm/internal/services/compute/parse"
-	"github.com/hashicorp/terraform-provider-azurerm/internal/tf/pluginsdk"
-	"github.com/hashicorp/terraform-provider-azurerm/utils"
 )
-
-type OrchestratedVirtualMachineScaleSetResource struct {
-}
-
-func TestAccOrchestratedVirtualMachineScaleSet_legacyBasicZonal(t *testing.T) {
-	data := acceptance.BuildTestData(t, "azurerm_orchestrated_virtual_machine_scale_set", "test")
-	r := OrchestratedVirtualMachineScaleSetResource{}
-
-	data.ResourceTest(t, r, []acceptance.TestStep{
-		{
-			Config: r.legacyBasic(data),
-			Check: acceptance.ComposeTestCheckFunc(
-				check.That(data.ResourceName).ExistsInAzure(r),
-			),
-		},
-		data.ImportStep("extensions_time_budget", "max_bid_price", "priority"),
-	})
-}
-
-func TestAccOrchestratedVirtualMachineScaleSet_legacyUpdateZonal(t *testing.T) {
-	data := acceptance.BuildTestData(t, "azurerm_orchestrated_virtual_machine_scale_set", "test")
-	r := OrchestratedVirtualMachineScaleSetResource{}
-
-	data.ResourceTest(t, r, []acceptance.TestStep{
-		{
-			Config: r.legacyBasic(data),
-			Check: acceptance.ComposeTestCheckFunc(
-				check.That(data.ResourceName).ExistsInAzure(r),
-			),
-		},
-		data.ImportStep("extensions_time_budget", "max_bid_price", "priority"),
-		{
-			Config: r.legacyUpdate(data),
-			Check: acceptance.ComposeTestCheckFunc(
-				check.That(data.ResourceName).ExistsInAzure(r),
-			),
-		},
-		data.ImportStep("extensions_time_budget", "max_bid_price", "priority"),
-		{
-			Config: r.legacyBasic(data),
-			Check: acceptance.ComposeTestCheckFunc(
-				check.That(data.ResourceName).ExistsInAzure(r),
-			),
-		},
-		data.ImportStep("extensions_time_budget", "max_bid_price", "priority"),
-	})
-}
-
-func TestAccOrchestratedVirtualMachineScaleSet_legacyBasicNonZonal(t *testing.T) {
-	data := acceptance.BuildTestData(t, "azurerm_orchestrated_virtual_machine_scale_set", "test")
-	r := OrchestratedVirtualMachineScaleSetResource{}
-
-	data.ResourceTest(t, r, []acceptance.TestStep{
-		{
-			Config: r.legacyBasicNonZonal(data),
-			Check: acceptance.ComposeTestCheckFunc(
-				check.That(data.ResourceName).ExistsInAzure(r),
-			),
-		},
-		data.ImportStep("extensions_time_budget", "max_bid_price", "priority"),
-	})
-}
-
-func TestAccOrchestratedVirtualMachineScaleSet_legacyRequiresImport(t *testing.T) {
-	data := acceptance.BuildTestData(t, "azurerm_orchestrated_virtual_machine_scale_set", "test")
-	r := OrchestratedVirtualMachineScaleSetResource{}
-
-	data.ResourceTest(t, r, []acceptance.TestStep{
-		{
-			Config: r.legacyBasic(data),
-			Check: acceptance.ComposeTestCheckFunc(
-				check.That(data.ResourceName).ExistsInAzure(r),
-			),
-		},
-		data.RequiresImportErrorStep(r.legacyRequiresImport),
-	})
-}
 
 func TestAccOrchestratedVirtualMachineScaleSet_basic(t *testing.T) {
 	data := acceptance.BuildTestData(t, "azurerm_orchestrated_virtual_machine_scale_set", "test")
@@ -173,85 +89,6 @@ func TestAccOrchestratedVirtualMachineScaleSet_withPPG(t *testing.T) {
 	})
 }
 
-func TestAccOrchestratedVirtualMachineScaleSet_basicPublicIP(t *testing.T) {
-	data := acceptance.BuildTestData(t, "azurerm_orchestrated_virtual_machine_scale_set", "test")
-	r := OrchestratedVirtualMachineScaleSetResource{}
-
-	data.ResourceTest(t, r, []acceptance.TestStep{
-		{
-			Config: r.basicPublicIP(data),
-			Check: acceptance.ComposeTestCheckFunc(
-				check.That(data.ResourceName).ExistsInAzure(r),
-			),
-		},
-		data.ImportStep("os_profile.0.linux_configuration.0.admin_password"),
-	})
-}
-
-func TestAccOrchestratedVirtualMachineScaleSet_basicPublicIP_simpleUpdate(t *testing.T) {
-	data := acceptance.BuildTestData(t, "azurerm_orchestrated_virtual_machine_scale_set", "test")
-	r := OrchestratedVirtualMachineScaleSetResource{}
-
-	data.ResourceTest(t, r, []acceptance.TestStep{
-		{
-			Config: r.basicEmptyPublicIP(data),
-			Check: acceptance.ComposeTestCheckFunc(
-				check.That(data.ResourceName).ExistsInAzure(r),
-			),
-		},
-		data.ImportStep("os_profile.0.linux_configuration.0.admin_password"),
-		{
-			Config: r.basicEmptyPublicIP_updated_tags(data),
-			Check: acceptance.ComposeTestCheckFunc(
-				check.That(data.ResourceName).ExistsInAzure(r),
-			),
-		},
-		data.ImportStep("os_profile.0.linux_configuration.0.admin_password"),
-	})
-}
-
-func TestAccOrchestratedVirtualMachineScaleSet_updateNetworkProfile(t *testing.T) {
-	data := acceptance.BuildTestData(t, "azurerm_orchestrated_virtual_machine_scale_set", "test")
-	r := OrchestratedVirtualMachineScaleSetResource{}
-
-	data.ResourceTest(t, r, []acceptance.TestStep{
-		{
-			Config: r.basicEmptyPublicIP(data),
-			Check: acceptance.ComposeTestCheckFunc(
-				check.That(data.ResourceName).ExistsInAzure(r),
-			),
-		},
-		data.ImportStep("os_profile.0.linux_configuration.0.admin_password"),
-		{
-			Config: r.basicEmptyNetworkProfile_true_ipforwarding(data),
-			Check:  acceptance.ComposeTestCheckFunc(),
-		},
-		data.ImportStep("os_profile.0.linux_configuration.0.admin_password"),
-	})
-}
-
-func TestAccOrchestratedVirtualMachineScaleSet_updateNetworkProfile_ipconfiguration_dns_name_label(t *testing.T) {
-	data := acceptance.BuildTestData(t, "azurerm_orchestrated_virtual_machine_scale_set", "test")
-	r := OrchestratedVirtualMachineScaleSetResource{}
-
-	data.ResourceTest(t, r, []acceptance.TestStep{
-		{
-			Config: r.basicEmptyPublicIP(data),
-			Check: acceptance.ComposeTestCheckFunc(
-				check.That(data.ResourceName).ExistsInAzure(r),
-			),
-		},
-		data.ImportStep("os_profile.0.linux_configuration.0.admin_password"),
-		{
-			Config: r.basicEmptyPublicIP_updatedDNS_label(data),
-			Check: acceptance.ComposeTestCheckFunc(
-				check.That(data.ResourceName).ExistsInAzure(r),
-			),
-		},
-		data.ImportStep("os_profile.0.linux_configuration.0.admin_password"),
-	})
-}
-
 func TestAccOrchestratedVirtualMachineScaleSet_verify_key_data_changed(t *testing.T) {
 	data := acceptance.BuildTestData(t, "azurerm_orchestrated_virtual_machine_scale_set", "test")
 	r := OrchestratedVirtualMachineScaleSetResource{}
@@ -289,51 +126,6 @@ func TestAccOrchestratedVirtualMachineScaleSet_basicApplicationSecurity(t *testi
 	})
 }
 
-func TestAccOrchestratedVirtualMachineScaleSet_basicAcceleratedNetworking(t *testing.T) {
-	data := acceptance.BuildTestData(t, "azurerm_orchestrated_virtual_machine_scale_set", "test")
-	r := OrchestratedVirtualMachineScaleSetResource{}
-
-	data.ResourceTest(t, r, []acceptance.TestStep{
-		{
-			Config: r.basicAcceleratedNetworking(data),
-			Check: acceptance.ComposeTestCheckFunc(
-				check.That(data.ResourceName).ExistsInAzure(r),
-			),
-		},
-		data.ImportStep("os_profile.0.linux_configuration.0.admin_password"),
-	})
-}
-
-func TestAccOrchestratedVirtualMachineScaleSet_basicIPForwarding(t *testing.T) {
-	data := acceptance.BuildTestData(t, "azurerm_orchestrated_virtual_machine_scale_set", "test")
-	r := OrchestratedVirtualMachineScaleSetResource{}
-
-	data.ResourceTest(t, r, []acceptance.TestStep{
-		{
-			Config: r.basicIPForwarding(data),
-			Check: acceptance.ComposeTestCheckFunc(
-				check.That(data.ResourceName).ExistsInAzure(r),
-			),
-		},
-		data.ImportStep("os_profile.0.linux_configuration.0.admin_password"),
-	})
-}
-
-func TestAccOrchestratedVirtualMachineScaleSet_basicDNSSettings(t *testing.T) {
-	data := acceptance.BuildTestData(t, "azurerm_orchestrated_virtual_machine_scale_set", "test")
-	r := OrchestratedVirtualMachineScaleSetResource{}
-
-	data.ResourceTest(t, r, []acceptance.TestStep{
-		{
-			Config: r.basicDNSSettings(data),
-			Check: acceptance.ComposeTestCheckFunc(
-				check.That(data.ResourceName).ExistsInAzure(r),
-			),
-		},
-		data.ImportStep("os_profile.0.linux_configuration.0.admin_password"),
-	})
-}
-
 func TestAccOrchestratedVirtualMachineScaleSet_bootDiagnostic(t *testing.T) {
 	data := acceptance.BuildTestData(t, "azurerm_orchestrated_virtual_machine_scale_set", "test")
 	r := OrchestratedVirtualMachineScaleSetResource{}
@@ -355,20 +147,6 @@ func TestAccOrchestratedVirtualMachineScaleSet_bootDiagnosticNoStorage(t *testin
 	data.ResourceTest(t, r, []acceptance.TestStep{
 		{
 			Config: r.bootDiagnostic_noStorage(data),
-			Check: acceptance.ComposeTestCheckFunc(
-				check.That(data.ResourceName).ExistsInAzure(r),
-			),
-		},
-	})
-}
-
-func TestAccOrchestratedVirtualMachineScaleSet_networkSecurityGroup(t *testing.T) {
-	data := acceptance.BuildTestData(t, "azurerm_orchestrated_virtual_machine_scale_set", "test")
-	r := OrchestratedVirtualMachineScaleSetResource{}
-
-	data.ResourceTest(t, r, []acceptance.TestStep{
-		{
-			Config: r.networkSecurityGroup(data),
 			Check: acceptance.ComposeTestCheckFunc(
 				check.That(data.ResourceName).ExistsInAzure(r),
 			),
@@ -427,21 +205,6 @@ func TestAccOrchestratedVirtualMachineScaleSet_customDataUpdated(t *testing.T) {
 				check.That(data.ResourceName).ExistsInAzure(r),
 			),
 		},
-	})
-}
-
-func TestAccOrchestratedVirtualMachineScaleSet_basicLinux_managedDisk(t *testing.T) {
-	data := acceptance.BuildTestData(t, "azurerm_orchestrated_virtual_machine_scale_set", "test")
-	r := OrchestratedVirtualMachineScaleSetResource{}
-
-	data.ResourceTest(t, r, []acceptance.TestStep{
-		{
-			Config: r.basicLinux_managedDisk(data),
-			Check: acceptance.ComposeTestCheckFunc(
-				check.That(data.ResourceName).ExistsInAzure(r),
-			),
-		},
-		data.ImportStep("os_profile.0.linux_configuration.0.admin_password"),
 	})
 }
 
@@ -523,163 +286,6 @@ func TestAccOrchestratedVirtualMachineScaleSet_applicationGateway(t *testing.T) 
 	})
 }
 
-func TestAccOrchestratedVirtualMachineScaleSet_loadBalancer(t *testing.T) {
-	data := acceptance.BuildTestData(t, "azurerm_orchestrated_virtual_machine_scale_set", "test")
-	r := OrchestratedVirtualMachineScaleSetResource{}
-
-	data.ResourceTest(t, r, []acceptance.TestStep{
-		{
-			Config: r.loadBalancerTemplate(data),
-			Check: acceptance.ComposeTestCheckFunc(
-				check.That(data.ResourceName).ExistsInAzure(r),
-				data.CheckWithClient(r.hasLoadBalancer),
-			),
-		},
-		data.ImportStep("os_profile.0.linux_configuration.0.admin_password"),
-	})
-}
-
-func TestAccOrchestratedVirtualMachineScaleSet_loadBalancerManagedDataDisks(t *testing.T) {
-	data := acceptance.BuildTestData(t, "azurerm_orchestrated_virtual_machine_scale_set", "test")
-	r := OrchestratedVirtualMachineScaleSetResource{}
-
-	data.ResourceTest(t, r, []acceptance.TestStep{
-		{
-			Config: r.loadBalancerTemplateManagedDataDisks(data),
-			Check: acceptance.ComposeTestCheckFunc(
-				check.That(data.ResourceName).ExistsInAzure(r),
-				check.That(data.ResourceName).Key("data_disk.#").HasValue("1"),
-			),
-		},
-	})
-}
-
-func TestAccOrchestratedVirtualMachineScaleSet_priority(t *testing.T) {
-	data := acceptance.BuildTestData(t, "azurerm_orchestrated_virtual_machine_scale_set", "test")
-	r := OrchestratedVirtualMachineScaleSetResource{}
-
-	data.ResourceTest(t, r, []acceptance.TestStep{
-		{
-			Config: r.priorityTemplate(data),
-			Check: acceptance.ComposeTestCheckFunc(
-				check.That(data.ResourceName).ExistsInAzure(r),
-				check.That(data.ResourceName).Key("priority").HasValue("Spot"),
-				check.That(data.ResourceName).Key("eviction_policy").HasValue("Deallocate"),
-			),
-		},
-	})
-}
-
-func TestAccOrchestratedVirtualMachineScaleSet_extensions(t *testing.T) {
-	data := acceptance.BuildTestData(t, "azurerm_orchestrated_virtual_machine_scale_set", "test")
-	r := OrchestratedVirtualMachineScaleSetResource{}
-
-	data.ResourceTest(t, r, []acceptance.TestStep{
-		{
-			Config: r.extensionTemplate(data),
-			Check: acceptance.ComposeTestCheckFunc(
-				check.That(data.ResourceName).ExistsInAzure(r),
-			),
-		},
-		data.ImportStep("os_profile.0.linux_configuration.0.admin_password", "extension.0.protected_settings"),
-	})
-}
-
-func TestAccOrchestratedVirtualMachineScaleSet_extensionsUpdate(t *testing.T) {
-	data := acceptance.BuildTestData(t, "azurerm_orchestrated_virtual_machine_scale_set", "test")
-	r := OrchestratedVirtualMachineScaleSetResource{}
-
-	data.ResourceTest(t, r, []acceptance.TestStep{
-		{
-			Config: r.extensionTemplate(data),
-			Check: acceptance.ComposeTestCheckFunc(
-				check.That(data.ResourceName).ExistsInAzure(r),
-			),
-		},
-		data.ImportStep("os_profile.0.linux_configuration.0.admin_password", "extension.0.protected_settings"),
-		{
-			Config: r.extensionTemplateUpdated(data),
-			Check: acceptance.ComposeTestCheckFunc(
-				check.That(data.ResourceName).ExistsInAzure(r),
-			),
-		},
-		data.ImportStep("os_profile.0.linux_configuration.0.admin_password", "extension.0.protected_settings"),
-	})
-}
-
-func TestAccOrchestratedVirtualMachineScaleSet_extensionsMultiple(t *testing.T) {
-	data := acceptance.BuildTestData(t, "azurerm_orchestrated_virtual_machine_scale_set", "test")
-	r := OrchestratedVirtualMachineScaleSetResource{}
-
-	data.ResourceTest(t, r, []acceptance.TestStep{
-		{
-			Config: r.multipleExtensionsTemplate(data),
-			Check: acceptance.ComposeTestCheckFunc(
-				check.That(data.ResourceName).ExistsInAzure(r),
-			),
-		},
-		data.ImportStep("os_profile.0.linux_configuration.0.admin_password", "extension.0.protected_settings"),
-	})
-}
-
-func TestAccOrchestratedVirtualMachineScaleSet_extensionsMultiple_on_existing_OVMSS(t *testing.T) {
-	data := acceptance.BuildTestData(t, "azurerm_orchestrated_virtual_machine_scale_set", "test")
-	r := OrchestratedVirtualMachineScaleSetResource{}
-
-	data.ResourceTest(t, r, []acceptance.TestStep{
-		{
-			Config: r.basic(data),
-			Check: acceptance.ComposeTestCheckFunc(
-				check.That(data.ResourceName).ExistsInAzure(r),
-			),
-		},
-		data.ImportStep("os_profile.0.linux_configuration.0.admin_password"),
-		{
-			Config: r.multipleExtensionsTemplate_provisionMultipleExtensionOnExistingOvmss(data),
-			Check: acceptance.ComposeTestCheckFunc(
-				check.That(data.ResourceName).ExistsInAzure(r),
-			),
-		},
-		data.ImportStep("os_profile.0.linux_configuration.0.admin_password", "extension.0.protected_settings"),
-	})
-}
-
-func TestAccOrchestratedVirtualMachineScaleSet_MSI_UserAssigned(t *testing.T) {
-	data := acceptance.BuildTestData(t, "azurerm_orchestrated_virtual_machine_scale_set", "test")
-	r := OrchestratedVirtualMachineScaleSetResource{}
-
-	data.ResourceTest(t, r, []acceptance.TestStep{
-		{
-			Config: r.userAssignedMSI(data),
-			Check: acceptance.ComposeTestCheckFunc(
-				check.That(data.ResourceName).ExistsInAzure(r),
-				check.That(data.ResourceName).Key("identity.0.type").HasValue("UserAssigned"),
-				check.That(data.ResourceName).Key("identity.0.identity_ids.#").HasValue("1"),
-			),
-		},
-		data.ImportStep("os_profile.0.linux_configuration.0.admin_password"),
-	})
-}
-
-func TestAccOrchestratedVirtualMachineScaleSet_NonStandardCasing(t *testing.T) {
-	data := acceptance.BuildTestData(t, "azurerm_orchestrated_virtual_machine_scale_set", "test")
-	r := OrchestratedVirtualMachineScaleSetResource{}
-
-	data.ResourceTest(t, r, []acceptance.TestStep{
-		{
-			Config: r.nonStandardCasing(data),
-			Check: acceptance.ComposeTestCheckFunc(
-				check.That(data.ResourceName).ExistsInAzure(r),
-			),
-		},
-		{
-			Config:             r.nonStandardCasing(data),
-			PlanOnly:           true,
-			ExpectNonEmptyPlan: false,
-		},
-	})
-}
-
 func TestAccOrchestratedVirtualMachineScaleSet_importLinux(t *testing.T) {
 	data := acceptance.BuildTestData(t, "azurerm_orchestrated_virtual_machine_scale_set", "test")
 	r := OrchestratedVirtualMachineScaleSetResource{}
@@ -695,20 +301,6 @@ func TestAccOrchestratedVirtualMachineScaleSet_importLinux(t *testing.T) {
 			"os_profile.0.linux_configuration.0.admin_password",
 			"os_profile.0.custom_data",
 		),
-	})
-}
-
-func TestAccOrchestratedVirtualMachineScaleSet_multipleNetworkProfiles(t *testing.T) {
-	data := acceptance.BuildTestData(t, "azurerm_orchestrated_virtual_machine_scale_set", "test")
-	r := OrchestratedVirtualMachineScaleSetResource{}
-
-	data.ResourceTest(t, r, []acceptance.TestStep{
-		{
-			Config: r.multipleNetworkProfiles(data),
-			Check: acceptance.ComposeTestCheckFunc(
-				check.That(data.ResourceName).ExistsInAzure(r),
-			),
-		},
 	})
 }
 
@@ -785,223 +377,6 @@ func TestAccOrchestratedVirtualMachineScaleSet_multipleNetworkProfiles(t *testin
 // 		},
 // 	})
 // }
-
-func TestAccOrchestratedVirtualMachineScaleSet_importBasic_managedDisk_withZones(t *testing.T) {
-	data := acceptance.BuildTestData(t, "azurerm_orchestrated_virtual_machine_scale_set", "test")
-	r := OrchestratedVirtualMachineScaleSetResource{}
-
-	data.ResourceTest(t, r, []acceptance.TestStep{
-		{
-			Config: r.basicLinux_managedDisk_withZones(data),
-			Check: acceptance.ComposeTestCheckFunc(
-				check.That(data.ResourceName).ExistsInAzure(r),
-			),
-		},
-		data.ImportStep("os_profile.0.linux_configuration.0.admin_password"),
-	})
-}
-
-func (t OrchestratedVirtualMachineScaleSetResource) Exists(ctx context.Context, clients *clients.Client, state *pluginsdk.InstanceState) (*bool, error) {
-	id, err := azure.ParseAzureResourceID(state.ID)
-	if err != nil {
-		return nil, err
-	}
-	resGroup := id.ResourceGroup
-	name := id.Path["virtualMachineScaleSets"]
-
-	// Upgrading to the 2021-07-01 exposed a new expand parameter in the GET method
-	resp, err := clients.Compute.VMScaleSetClient.Get(ctx, resGroup, name, compute.ExpandTypesForGetVMScaleSetsUserData)
-	if err != nil {
-		return nil, fmt.Errorf("retrieving Virtual Machine Scale Set %q", id)
-	}
-
-	return utils.Bool(resp.ID != nil), nil
-}
-
-func (OrchestratedVirtualMachineScaleSetResource) Destroy(ctx context.Context, client *clients.Client, state *pluginsdk.InstanceState) (*bool, error) {
-	id, err := parse.VirtualMachineScaleSetID(state.ID)
-	if err != nil {
-		return nil, err
-	}
-
-	// this is a preview feature we don't want to use right now
-	var forceDelete *bool = nil
-	future, err := client.Compute.VMScaleSetClient.Delete(ctx, id.ResourceGroup, id.Name, forceDelete)
-	if err != nil {
-		return nil, fmt.Errorf("Bad: deleting %s: %+v", *id, err)
-	}
-
-	if err = future.WaitForCompletionRef(ctx, client.Compute.VMScaleSetClient.Client); err != nil {
-		return nil, fmt.Errorf("Bad: waiting for deletion of %s: %+v", *id, err)
-	}
-
-	return utils.Bool(true), nil
-}
-
-func (OrchestratedVirtualMachineScaleSetResource) hasLoadBalancer(ctx context.Context, client *clients.Client, state *pluginsdk.InstanceState) error {
-	id, err := parse.VirtualMachineScaleSetID(state.ID)
-	if err != nil {
-		return err
-	}
-
-	// Upgrading to the 2021-07-01 exposed a new expand parameter in the GET method
-	read, err := client.Compute.VMScaleSetClient.Get(ctx, id.ResourceGroup, id.Name, compute.ExpandTypesForGetVMScaleSetsUserData)
-	if err != nil {
-		return err
-	}
-
-	if props := read.VirtualMachineScaleSetProperties; props != nil {
-		if vmProfile := props.VirtualMachineProfile; vmProfile != nil {
-			if nwProfile := vmProfile.NetworkProfile; nwProfile != nil {
-				if nics := nwProfile.NetworkInterfaceConfigurations; nics != nil {
-					for _, nic := range *nics {
-						if nic.IPConfigurations == nil {
-							continue
-						}
-
-						for _, config := range *nic.IPConfigurations {
-							if config.LoadBalancerBackendAddressPools == nil {
-								continue
-							}
-
-							if len(*config.LoadBalancerBackendAddressPools) > 0 {
-								return nil
-							}
-						}
-					}
-				}
-			}
-		}
-	}
-
-	return fmt.Errorf("load balancer configuration was missing")
-}
-
-func (OrchestratedVirtualMachineScaleSetResource) hasApplicationGateway(ctx context.Context, client *clients.Client, state *pluginsdk.InstanceState) error {
-	id, err := parse.VirtualMachineScaleSetID(state.ID)
-	if err != nil {
-		return err
-	}
-
-	// Upgrading to the 2021-07-01 exposed a new expand parameter in the GET method
-	read, err := client.Compute.VMScaleSetClient.Get(ctx, id.ResourceGroup, id.Name, compute.ExpandTypesForGetVMScaleSetsUserData)
-	if err != nil {
-		return err
-	}
-
-	if props := read.VirtualMachineScaleSetProperties; props != nil {
-		if vmProfile := props.VirtualMachineProfile; vmProfile != nil {
-			if nwProfile := vmProfile.NetworkProfile; nwProfile != nil {
-				if nics := nwProfile.NetworkInterfaceConfigurations; nics != nil {
-					for _, nic := range *nics {
-						if nic.IPConfigurations == nil {
-							continue
-						}
-
-						for _, config := range *nic.IPConfigurations {
-							if config.ApplicationGatewayBackendAddressPools == nil {
-								continue
-							}
-
-							if len(*config.ApplicationGatewayBackendAddressPools) > 0 {
-								return nil
-							}
-						}
-					}
-				}
-			}
-		}
-	}
-
-	return fmt.Errorf("application gateway configuration was missing")
-}
-
-// legacy test cases for backward compatibility validation
-func (r OrchestratedVirtualMachineScaleSetResource) legacyBasic(data acceptance.TestData) string {
-	return fmt.Sprintf(`
-%s
-
-resource "azurerm_orchestrated_virtual_machine_scale_set" "test" {
-  name                = "acctestOVMSS-%d"
-  location            = azurerm_resource_group.test.location
-  resource_group_name = azurerm_resource_group.test.name
-
-  platform_fault_domain_count = 1
-
-  zones = ["1"]
-
-  tags = {
-    ENV = "Test"
-  }
-}
-`, r.legacyTemplate(data), data.RandomInteger)
-}
-
-func (r OrchestratedVirtualMachineScaleSetResource) legacyUpdate(data acceptance.TestData) string {
-	return fmt.Sprintf(`
-%s
-
-resource "azurerm_orchestrated_virtual_machine_scale_set" "test" {
-  name                = "acctestOVMSS-%d"
-  location            = azurerm_resource_group.test.location
-  resource_group_name = azurerm_resource_group.test.name
-
-  platform_fault_domain_count = 1
-
-  zones = ["1"]
-
-  tags = {
-    ENV = "Test",
-    FOO = "Bar"
-  }
-}
-`, r.legacyTemplate(data), data.RandomInteger)
-}
-
-func (r OrchestratedVirtualMachineScaleSetResource) legacyRequiresImport(data acceptance.TestData) string {
-	return fmt.Sprintf(`
-%s
-
-resource "azurerm_orchestrated_virtual_machine_scale_set" "import" {
-  name                = azurerm_orchestrated_virtual_machine_scale_set.test.name
-  location            = azurerm_orchestrated_virtual_machine_scale_set.test.location
-  resource_group_name = azurerm_orchestrated_virtual_machine_scale_set.test.resource_group_name
-
-  platform_fault_domain_count = azurerm_orchestrated_virtual_machine_scale_set.test.platform_fault_domain_count
-}
-`, r.legacyBasic(data))
-}
-
-func (r OrchestratedVirtualMachineScaleSetResource) legacyBasicNonZonal(data acceptance.TestData) string {
-	return fmt.Sprintf(`
-%s
-
-resource "azurerm_orchestrated_virtual_machine_scale_set" "test" {
-  name                = "acctestOVMSS-%d"
-  location            = azurerm_resource_group.test.location
-  resource_group_name = azurerm_resource_group.test.name
-
-  platform_fault_domain_count = 2
-
-  tags = {
-    ENV = "Test"
-  }
-}
-`, r.legacyTemplate(data), data.RandomInteger)
-}
-
-func (OrchestratedVirtualMachineScaleSetResource) legacyTemplate(data acceptance.TestData) string {
-	return fmt.Sprintf(`
-provider "azurerm" {
-  features {}
-}
-
-resource "azurerm_resource_group" "test" {
-  name     = "acctestRG-OVMSS-%d"
-  location = "%s"
-}
-`, data.RandomInteger, data.Locations.Primary)
-}
 
 // Net new tests for the 2021-07-01 version of the API
 func (OrchestratedVirtualMachineScaleSetResource) basic(data acceptance.TestData) string {
@@ -1120,86 +495,6 @@ resource "azurerm_orchestrated_virtual_machine_scale_set" "import" {
   }
 }
 `, r.basic(data), data.RandomInteger, data.RandomInteger)
-}
-
-func (OrchestratedVirtualMachineScaleSetResource) userAssignedMSI(data acceptance.TestData) string {
-	r := OrchestratedVirtualMachineScaleSetResource{}
-	return fmt.Sprintf(`
-provider "azurerm" {
-  features {}
-}
-
-resource "azurerm_resource_group" "test" {
-  name     = "acctestRG-OVMSS-%[1]d"
-  location = "%[2]s"
-}
-
-%[3]s
-
-resource "azurerm_user_assigned_identity" "test" {
-  resource_group_name = azurerm_resource_group.test.name
-  location            = azurerm_resource_group.test.location
-
-  name = "acctest%[4]s"
-}
-
-resource "azurerm_orchestrated_virtual_machine_scale_set" "test" {
-  name                = "acctestOVMSS-%[1]d"
-  location            = azurerm_resource_group.test.location
-  resource_group_name = azurerm_resource_group.test.name
-
-  zones = []
-
-  sku_name  = "Standard_D1_v2"
-  instances = 1
-
-  identity {
-    type         = "UserAssigned"
-    identity_ids = [azurerm_user_assigned_identity.test.id]
-  }
-
-  platform_fault_domain_count = 2
-
-  os_profile {
-    linux_configuration {
-      computer_name_prefix = "testvm-%[1]d"
-      admin_username       = "myadmin"
-      admin_password       = "Passwword1234"
-
-      disable_password_authentication = false
-    }
-  }
-
-  network_interface {
-    name    = "TestNetworkProfile-%[1]d"
-    primary = true
-
-    ip_configuration {
-      name      = "TestIPConfiguration"
-      primary   = true
-      subnet_id = azurerm_subnet.test.id
-
-      public_ip_address {
-        name                    = "TestPublicIPConfiguration"
-        domain_name_label       = "test-domain-label"
-        idle_timeout_in_minutes = 4
-      }
-    }
-  }
-
-  os_disk {
-    storage_account_type = "Standard_LRS"
-    caching              = "ReadWrite"
-  }
-
-  source_image_reference {
-    publisher = "Canonical"
-    offer     = "UbuntuServer"
-    sku       = "16.04-LTS"
-    version   = "latest"
-  }
-}
-`, data.RandomInteger, data.Locations.Primary, r.natgateway_template(data), data.RandomString)
 }
 
 func (OrchestratedVirtualMachineScaleSetResource) evictionPolicyDelete(data acceptance.TestData) string {
@@ -1345,353 +640,6 @@ resource "azurerm_orchestrated_virtual_machine_scale_set" "test" {
 `, data.RandomInteger, data.Locations.Primary, r.natgateway_template(data))
 }
 
-func (OrchestratedVirtualMachineScaleSetResource) basicPublicIP(data acceptance.TestData) string {
-	r := OrchestratedVirtualMachineScaleSetResource{}
-	return fmt.Sprintf(`
-provider "azurerm" {
-  features {}
-}
-
-resource "azurerm_resource_group" "test" {
-  name     = "acctestRG-OVMSS-%[1]d"
-  location = "%[2]s"
-}
-
-%[3]s
-
-resource "azurerm_orchestrated_virtual_machine_scale_set" "test" {
-  name                = "acctestOVMSS-%[1]d"
-  location            = azurerm_resource_group.test.location
-  resource_group_name = azurerm_resource_group.test.name
-
-  sku_name  = "Standard_D1_v2"
-  instances = 2
-
-  platform_fault_domain_count = 2
-
-  os_profile {
-    linux_configuration {
-      computer_name_prefix = "testvm-%[1]d"
-      admin_username       = "myadmin"
-      admin_password       = "Passwword1234"
-
-      disable_password_authentication = false
-    }
-  }
-
-  network_interface {
-    name    = "TestNetworkProfile-%[1]d"
-    primary = true
-
-    ip_configuration {
-      name      = "TestIPConfiguration"
-      subnet_id = azurerm_subnet.test.id
-      primary   = true
-
-      public_ip_address {
-        name                    = "TestPublicIPConfiguration"
-        domain_name_label       = "test-domain-label"
-        idle_timeout_in_minutes = 4
-      }
-    }
-  }
-
-  os_disk {
-    storage_account_type = "Standard_LRS"
-    caching              = "ReadWrite"
-  }
-
-  source_image_reference {
-    publisher = "Canonical"
-    offer     = "UbuntuServer"
-    sku       = "16.04-LTS"
-    version   = "latest"
-  }
-}
-`, data.RandomInteger, data.Locations.Primary, r.natgateway_template(data))
-}
-
-func (OrchestratedVirtualMachineScaleSetResource) basicEmptyPublicIP(data acceptance.TestData) string {
-	r := OrchestratedVirtualMachineScaleSetResource{}
-	return fmt.Sprintf(`
-provider "azurerm" {
-  features {}
-}
-
-resource "azurerm_resource_group" "test" {
-  name     = "acctestRG-OVMSS-%[1]d"
-  location = "%[2]s"
-}
-
-%[3]s
-
-resource "azurerm_orchestrated_virtual_machine_scale_set" "test" {
-  name                = "acctestOVMSS-%[1]d"
-  location            = azurerm_resource_group.test.location
-  resource_group_name = azurerm_resource_group.test.name
-
-  tags = {
-    state = "create"
-  }
-
-  sku_name  = "Standard_D1_v2"
-  instances = 0
-
-  platform_fault_domain_count = 2
-
-  os_profile {
-    linux_configuration {
-      computer_name_prefix = "testvm-%[1]d"
-      admin_username       = "myadmin"
-      admin_password       = "Passwword1234"
-
-      disable_password_authentication = false
-    }
-  }
-
-  network_interface {
-    name    = "TestNetworkProfile-%[1]d"
-    primary = true
-
-    ip_configuration {
-      name      = "TestIPConfiguration"
-      subnet_id = azurerm_subnet.test.id
-      primary   = true
-
-      public_ip_address {
-        name                    = "TestPublicIPConfiguration"
-        domain_name_label       = "test-domain-label"
-        idle_timeout_in_minutes = 4
-      }
-    }
-  }
-
-  os_disk {
-    storage_account_type = "Standard_LRS"
-    caching              = "ReadWrite"
-  }
-
-  source_image_reference {
-    publisher = "Canonical"
-    offer     = "UbuntuServer"
-    sku       = "16.04-LTS"
-    version   = "latest"
-  }
-}
-`, data.RandomInteger, data.Locations.Primary, r.natgateway_template(data))
-}
-
-func (OrchestratedVirtualMachineScaleSetResource) basicEmptyPublicIP_updated_tags(data acceptance.TestData) string {
-	r := OrchestratedVirtualMachineScaleSetResource{}
-	return fmt.Sprintf(`
-provider "azurerm" {
-  features {}
-}
-
-resource "azurerm_resource_group" "test" {
-  name     = "acctestRG-OVMSS-%[1]d"
-  location = "%[2]s"
-}
-
-%[3]s
-
-resource "azurerm_orchestrated_virtual_machine_scale_set" "test" {
-  name                = "acctestOVMSS-%[1]d"
-  location            = azurerm_resource_group.test.location
-  resource_group_name = azurerm_resource_group.test.name
-
-  tags = {
-    state = "update"
-  }
-
-  sku_name  = "Standard_D1_v2"
-  instances = 0
-
-  platform_fault_domain_count = 2
-
-  os_profile {
-    linux_configuration {
-      computer_name_prefix = "testvm-%[1]d"
-      admin_username       = "myadmin"
-      admin_password       = "Passwword1234"
-
-      disable_password_authentication = false
-    }
-  }
-
-  network_interface {
-    name    = "TestNetworkProfile-%[1]d"
-    primary = true
-
-    ip_configuration {
-      name      = "TestIPConfiguration"
-      subnet_id = azurerm_subnet.test.id
-      primary   = true
-
-      public_ip_address {
-        name                    = "TestPublicIPConfiguration"
-        domain_name_label       = "test-domain-label"
-        idle_timeout_in_minutes = 4
-      }
-    }
-  }
-
-  os_disk {
-    storage_account_type = "Standard_LRS"
-    caching              = "ReadWrite"
-  }
-
-  source_image_reference {
-    publisher = "Canonical"
-    offer     = "UbuntuServer"
-    sku       = "16.04-LTS"
-    version   = "latest"
-  }
-}
-`, data.RandomInteger, data.Locations.Primary, r.natgateway_template(data))
-}
-
-func (OrchestratedVirtualMachineScaleSetResource) basicEmptyNetworkProfile_true_ipforwarding(data acceptance.TestData) string {
-	r := OrchestratedVirtualMachineScaleSetResource{}
-	return fmt.Sprintf(`
-provider "azurerm" {
-  features {}
-}
-
-resource "azurerm_resource_group" "test" {
-  name     = "acctestRG-OVMSS-%[1]d"
-  location = "%[2]s"
-}
-
-%[3]s
-
-resource "azurerm_orchestrated_virtual_machine_scale_set" "test" {
-  name                = "acctestOVMSS-%[1]d"
-  location            = azurerm_resource_group.test.location
-  resource_group_name = azurerm_resource_group.test.name
-
-  tags = {
-    state = "update"
-  }
-
-  sku_name  = "Standard_D1_v2"
-  instances = 0
-
-  platform_fault_domain_count = 2
-
-  os_profile {
-    linux_configuration {
-      computer_name_prefix = "testvm-%[1]d"
-      admin_username       = "myadmin"
-      admin_password       = "Passwword1234"
-
-      disable_password_authentication = false
-    }
-  }
-
-  network_interface {
-    name                 = "TestNetworkProfile-%[1]d"
-    primary              = true
-    enable_ip_forwarding = true
-
-    ip_configuration {
-      name      = "TestIPConfiguration"
-      subnet_id = azurerm_subnet.test.id
-      primary   = true
-
-      public_ip_address {
-        name                    = "TestPublicIPConfiguration"
-        domain_name_label       = "test-domain-label"
-        idle_timeout_in_minutes = 4
-      }
-    }
-  }
-
-  os_disk {
-    storage_account_type = "Standard_LRS"
-    caching              = "ReadWrite"
-  }
-
-  source_image_reference {
-    publisher = "Canonical"
-    offer     = "UbuntuServer"
-    sku       = "16.04-LTS"
-    version   = "latest"
-  }
-}
-`, data.RandomInteger, data.Locations.Primary, r.natgateway_template(data))
-}
-
-func (OrchestratedVirtualMachineScaleSetResource) basicEmptyPublicIP_updatedDNS_label(data acceptance.TestData) string {
-	r := OrchestratedVirtualMachineScaleSetResource{}
-	return fmt.Sprintf(`
-provider "azurerm" {
-  features {}
-}
-
-resource "azurerm_resource_group" "test" {
-  name     = "acctestRG-OVMSS-%[1]d"
-  location = "%[2]s"
-}
-
-%[3]s
-
-resource "azurerm_orchestrated_virtual_machine_scale_set" "test" {
-  name                = "acctestOVMSS-%[1]d"
-  location            = azurerm_resource_group.test.location
-  resource_group_name = azurerm_resource_group.test.name
-
-  tags = {
-    state = "create"
-  }
-
-  sku_name  = "Standard_D1_v2"
-  instances = 0
-
-  platform_fault_domain_count = 2
-
-  os_profile {
-    linux_configuration {
-      computer_name_prefix = "testvm-%[1]d"
-      admin_username       = "myadmin"
-      admin_password       = "Passwword1234"
-
-      disable_password_authentication = false
-    }
-  }
-
-  network_interface {
-    name    = "TestNetworkProfile-%[1]d"
-    primary = true
-
-    ip_configuration {
-      name      = "TestIPConfiguration"
-      subnet_id = azurerm_subnet.test.id
-      primary   = true
-
-      public_ip_address {
-        name                    = "TestPublicIPConfiguration"
-        domain_name_label       = "updated-domain-label"
-        idle_timeout_in_minutes = 4
-      }
-    }
-  }
-
-  os_disk {
-    storage_account_type = "Standard_LRS"
-    caching              = "ReadWrite"
-  }
-
-  source_image_reference {
-    publisher = "Canonical"
-    offer     = "UbuntuServer"
-    sku       = "16.04-LTS"
-    version   = "latest"
-  }
-}
-`, data.RandomInteger, data.Locations.Primary, r.natgateway_template(data))
-}
-
 func (OrchestratedVirtualMachineScaleSetResource) basicApplicationSecurity(data acceptance.TestData) string {
 	r := OrchestratedVirtualMachineScaleSetResource{}
 	return fmt.Sprintf(`
@@ -1748,281 +696,6 @@ resource "azurerm_orchestrated_virtual_machine_scale_set" "test" {
       }
 
       application_security_group_ids = [azurerm_application_security_group.test.id]
-    }
-  }
-
-  os_disk {
-    storage_account_type = "Standard_LRS"
-    caching              = "ReadWrite"
-  }
-
-  source_image_reference {
-    publisher = "Canonical"
-    offer     = "UbuntuServer"
-    sku       = "16.04-LTS"
-    version   = "latest"
-  }
-}
-`, data.RandomInteger, data.Locations.Primary, r.natgateway_template(data))
-}
-
-func (OrchestratedVirtualMachineScaleSetResource) basicAcceleratedNetworking(data acceptance.TestData) string {
-	r := OrchestratedVirtualMachineScaleSetResource{}
-	return fmt.Sprintf(`
-provider "azurerm" {
-  features {}
-}
-
-resource "azurerm_resource_group" "test" {
-  name     = "acctestRG-OVMSS-%[1]d"
-  location = "%[2]s"
-}
-
-%[3]s
-
-resource "azurerm_orchestrated_virtual_machine_scale_set" "test" {
-  name                = "acctestOVMSS-%[1]d"
-  location            = azurerm_resource_group.test.location
-  resource_group_name = azurerm_resource_group.test.name
-
-  sku_name  = "Standard_D4_v2"
-  instances = 2
-
-  platform_fault_domain_count = 2
-
-  os_profile {
-    linux_configuration {
-      computer_name_prefix = "testvm-%[1]d"
-      admin_username       = "myadmin"
-      admin_password       = "Passwword1234"
-
-      disable_password_authentication = false
-    }
-  }
-
-  network_interface {
-    name                          = "TestNetworkProfile-%[1]d"
-    primary                       = true
-    enable_accelerated_networking = true
-
-    ip_configuration {
-      name      = "TestIPConfiguration"
-      primary   = true
-      subnet_id = azurerm_subnet.test.id
-
-      public_ip_address {
-        name                    = "TestPublicIPConfiguration"
-        domain_name_label       = "test-domain-label"
-        idle_timeout_in_minutes = 4
-      }
-    }
-  }
-
-  os_disk {
-    storage_account_type = "Standard_LRS"
-    caching              = "ReadWrite"
-  }
-
-  source_image_reference {
-    publisher = "Canonical"
-    offer     = "UbuntuServer"
-    sku       = "16.04-LTS"
-    version   = "latest"
-  }
-}
-`, data.RandomInteger, data.Locations.Primary, r.natgateway_template(data))
-}
-
-func (OrchestratedVirtualMachineScaleSetResource) basicIPForwarding(data acceptance.TestData) string {
-	r := OrchestratedVirtualMachineScaleSetResource{}
-	return fmt.Sprintf(`
-provider "azurerm" {
-  features {}
-}
-
-resource "azurerm_resource_group" "test" {
-  name     = "acctestRG-OVMSS-%[1]d"
-  location = "%[2]s"
-}
-
-%[3]s
-
-resource "azurerm_orchestrated_virtual_machine_scale_set" "test" {
-  name                = "acctestOVMSS-%[1]d"
-  location            = azurerm_resource_group.test.location
-  resource_group_name = azurerm_resource_group.test.name
-
-  sku_name  = "Standard_D4_v2"
-  instances = 2
-
-  platform_fault_domain_count = 2
-
-  os_profile {
-    linux_configuration {
-      computer_name_prefix = "testvm-%[1]d"
-      admin_username       = "myadmin"
-      admin_password       = "Passwword1234"
-
-      disable_password_authentication = false
-    }
-  }
-
-  network_interface {
-    name                 = "TestNetworkProfile-%[1]d"
-    primary              = true
-    enable_ip_forwarding = true
-
-    ip_configuration {
-      name      = "TestIPConfiguration"
-      primary   = true
-      subnet_id = azurerm_subnet.test.id
-
-      public_ip_address {
-        name                    = "TestPublicIPConfiguration"
-        domain_name_label       = "test-domain-label"
-        idle_timeout_in_minutes = 4
-      }
-    }
-  }
-
-  os_disk {
-    storage_account_type = "Standard_LRS"
-    caching              = "ReadWrite"
-  }
-
-  source_image_reference {
-    publisher = "Canonical"
-    offer     = "UbuntuServer"
-    sku       = "16.04-LTS"
-    version   = "latest"
-  }
-}
-`, data.RandomInteger, data.Locations.Primary, r.natgateway_template(data))
-}
-
-func (OrchestratedVirtualMachineScaleSetResource) basicDNSSettings(data acceptance.TestData) string {
-	r := OrchestratedVirtualMachineScaleSetResource{}
-	return fmt.Sprintf(`
-provider "azurerm" {
-  features {}
-}
-
-resource "azurerm_resource_group" "test" {
-  name     = "acctestRG-OVMSS-%[1]d"
-  location = "%[2]s"
-}
-
-%[3]s
-
-resource "azurerm_orchestrated_virtual_machine_scale_set" "test" {
-  name                = "acctestOVMSS-%[1]d"
-  location            = azurerm_resource_group.test.location
-  resource_group_name = azurerm_resource_group.test.name
-
-  sku_name  = "Standard_D4_v2"
-  instances = 2
-
-  platform_fault_domain_count = 2
-
-  os_profile {
-    linux_configuration {
-      computer_name_prefix = "testvm-%[1]d"
-      admin_username       = "myadmin"
-      admin_password       = "Passwword1234"
-
-      disable_password_authentication = false
-    }
-  }
-
-  network_interface {
-    name    = "TestNetworkProfile-%[1]d"
-    primary = true
-
-    dns_servers = ["8.8.8.8", "8.8.4.4"]
-
-    ip_configuration {
-      name      = "TestIPConfiguration"
-      primary   = true
-      subnet_id = azurerm_subnet.test.id
-
-      public_ip_address {
-        name                    = "TestPublicIPConfiguration"
-        domain_name_label       = "test-domain-label"
-        idle_timeout_in_minutes = 4
-      }
-    }
-  }
-
-  os_disk {
-    storage_account_type = "Standard_LRS"
-    caching              = "ReadWrite"
-  }
-
-  source_image_reference {
-    publisher = "Canonical"
-    offer     = "UbuntuServer"
-    sku       = "16.04-LTS"
-    version   = "latest"
-  }
-}
-`, data.RandomInteger, data.Locations.Primary, r.natgateway_template(data))
-}
-
-func (OrchestratedVirtualMachineScaleSetResource) networkSecurityGroup(data acceptance.TestData) string {
-	r := OrchestratedVirtualMachineScaleSetResource{}
-	return fmt.Sprintf(`
-provider "azurerm" {
-  features {}
-}
-
-resource "azurerm_resource_group" "test" {
-  name     = "acctestRG-OVMSS-%[1]d"
-  location = "%[2]s"
-}
-
-%[3]s
-
-resource "azurerm_network_security_group" "test" {
-  name                = "acceptanceTestSecurityGroup-%[1]d"
-  location            = azurerm_resource_group.test.location
-  resource_group_name = azurerm_resource_group.test.name
-}
-
-resource "azurerm_orchestrated_virtual_machine_scale_set" "test" {
-  name                = "acctestOVMSS-%[1]d"
-  location            = azurerm_resource_group.test.location
-  resource_group_name = azurerm_resource_group.test.name
-
-  sku_name  = "Standard_D1_v2"
-  instances = 2
-
-  platform_fault_domain_count = 2
-
-  os_profile {
-    linux_configuration {
-      computer_name_prefix = "testvm-%[1]d"
-      admin_username       = "myadmin"
-      admin_password       = "Passwword1234"
-
-      disable_password_authentication = false
-    }
-  }
-
-  network_interface {
-    name                      = "TestNetworkProfile-%[1]d"
-    primary                   = true
-    network_security_group_id = azurerm_network_security_group.test.id
-
-    ip_configuration {
-      name      = "TestIPConfiguration"
-      subnet_id = azurerm_subnet.test.id
-      primary   = true
-
-      public_ip_address {
-        name                    = "TestPublicIPConfiguration"
-        domain_name_label       = "test-domain-label"
-        idle_timeout_in_minutes = 4
-      }
     }
   }
 
@@ -2685,139 +1358,6 @@ resource "azurerm_orchestrated_virtual_machine_scale_set" "test" {
 `, data.RandomInteger, data.Locations.Primary)
 }
 
-func (OrchestratedVirtualMachineScaleSetResource) basicLinux_managedDisk(data acceptance.TestData) string {
-	r := OrchestratedVirtualMachineScaleSetResource{}
-	return fmt.Sprintf(`
-provider "azurerm" {
-  features {}
-}
-
-resource "azurerm_resource_group" "test" {
-  name     = "acctestRG-OVMSS-%[1]d"
-  location = "%[2]s"
-}
-
-%[3]s
-
-resource "azurerm_orchestrated_virtual_machine_scale_set" "test" {
-  name                = "acctestOVMSS-%[1]d"
-  location            = azurerm_resource_group.test.location
-  resource_group_name = azurerm_resource_group.test.name
-
-  sku_name  = "Standard_D1_v2"
-  instances = 2
-
-  platform_fault_domain_count = 2
-
-  os_profile {
-    linux_configuration {
-      computer_name_prefix = "testvm-%[1]d"
-      admin_username       = "myadmin"
-      admin_password       = "Passwword1234"
-
-      disable_password_authentication = false
-    }
-  }
-
-  network_interface {
-    name    = "TestNetworkProfile-%[1]d"
-    primary = true
-
-    ip_configuration {
-      name      = "TestIPConfiguration"
-      primary   = true
-      subnet_id = azurerm_subnet.test.id
-
-      public_ip_address {
-        name                    = "TestPublicIPConfiguration"
-        domain_name_label       = "test-domain-label"
-        idle_timeout_in_minutes = 4
-      }
-    }
-  }
-
-  os_disk {
-    storage_account_type = "Standard_LRS"
-    caching              = "ReadWrite"
-  }
-
-  source_image_reference {
-    publisher = "Canonical"
-    offer     = "UbuntuServer"
-    sku       = "16.04-LTS"
-    version   = "latest"
-  }
-}
-`, data.RandomInteger, data.Locations.Primary, r.natgateway_template(data))
-}
-
-func (OrchestratedVirtualMachineScaleSetResource) basicLinux_managedDisk_withZones(data acceptance.TestData) string {
-	r := OrchestratedVirtualMachineScaleSetResource{}
-	return fmt.Sprintf(`
-provider "azurerm" {
-  features {}
-}
-
-resource "azurerm_resource_group" "test" {
-  name     = "acctestRG-OVMSS-%[1]d"
-  location = "%[2]s"
-}
-
-%[3]s
-
-resource "azurerm_orchestrated_virtual_machine_scale_set" "test" {
-  name                = "acctestOVMSS-%[1]d"
-  location            = azurerm_resource_group.test.location
-  resource_group_name = azurerm_resource_group.test.name
-
-  zones                       = ["1"]
-  platform_fault_domain_count = 1
-
-  sku_name  = "Standard_D1_v2"
-  instances = 2
-
-  os_profile {
-    linux_configuration {
-      computer_name_prefix = "testvm-%[1]d"
-      admin_username       = "myadmin"
-      admin_password       = "Passwword1234"
-
-      disable_password_authentication = false
-    }
-  }
-
-  network_interface {
-    name    = "TestNetworkProfile-%[1]d"
-    primary = true
-
-    ip_configuration {
-      name      = "TestIPConfiguration"
-      primary   = true
-      subnet_id = azurerm_subnet.test.id
-
-      public_ip_address {
-        name                    = "TestPublicIPConfiguration"
-        domain_name_label       = "test-domain-label"
-        idle_timeout_in_minutes = 4
-      }
-    }
-  }
-
-  os_disk {
-    storage_account_type = "Standard_LRS"
-    caching              = "ReadWrite"
-  }
-
-  source_image_reference {
-    publisher = "Canonical"
-    offer     = "UbuntuServer"
-    sku       = "16.04-LTS"
-    version   = "latest"
-  }
-}
-`, data.RandomInteger, data.Locations.Primary, r.natgateway_template(data))
-}
-
 func (OrchestratedVirtualMachineScaleSetResource) applicationGatewayTemplate(data acceptance.TestData) string {
 	r := OrchestratedVirtualMachineScaleSetResource{}
 	return fmt.Sprintf(`
@@ -2967,761 +1507,6 @@ resource "azurerm_application_gateway" "test" {
 
   tags = {
     environment = "tf01"
-  }
-}
-`, data.RandomInteger, data.Locations.Primary, r.natgateway_template(data))
-}
-
-func (OrchestratedVirtualMachineScaleSetResource) loadBalancerTemplate(data acceptance.TestData) string {
-	return fmt.Sprintf(`
-provider "azurerm" {
-  features {}
-}
-
-resource "azurerm_resource_group" "test" {
-  name     = "acctestRG-OVMSS-%[1]d"
-  location = "%[2]s"
-}
-
-resource "azurerm_virtual_network" "test" {
-  name                = "acctvn-%[1]d"
-  address_space       = ["10.0.0.0/16"]
-  location            = azurerm_resource_group.test.location
-  resource_group_name = azurerm_resource_group.test.name
-}
-
-resource "azurerm_subnet" "test" {
-  name                 = "acctsub-%[1]d"
-  resource_group_name  = azurerm_resource_group.test.name
-  virtual_network_name = azurerm_virtual_network.test.name
-  address_prefixes     = ["10.0.2.0/24"]
-}
-
-resource "azurerm_lb" "test" {
-  name                = "acctestlb-%[1]d"
-  location            = azurerm_resource_group.test.location
-  resource_group_name = azurerm_resource_group.test.name
-
-  sku = "Standard"
-
-  frontend_ip_configuration {
-    name                          = "default"
-    subnet_id                     = azurerm_subnet.test.id
-    private_ip_address_allocation = "Dynamic"
-  }
-}
-
-resource "azurerm_lb_backend_address_pool" "test" {
-  name            = "test"
-  loadbalancer_id = azurerm_lb.test.id
-}
-
-resource "azurerm_orchestrated_virtual_machine_scale_set" "test" {
-  name                = "acctestOVMSS-%[1]d"
-  location            = azurerm_resource_group.test.location
-  resource_group_name = azurerm_resource_group.test.name
-
-  sku_name  = "Standard_D1_v2"
-  instances = 1
-
-  platform_fault_domain_count = 2
-
-  os_profile {
-    linux_configuration {
-      computer_name_prefix = "testvm-%[1]d"
-      admin_username       = "myadmin"
-      admin_password       = "Passwword1234"
-
-      disable_password_authentication = false
-    }
-  }
-
-  network_interface {
-    name    = "TestNetworkProfile"
-    primary = true
-
-    ip_configuration {
-      name      = "TestIPConfiguration"
-      primary   = true
-      subnet_id = azurerm_subnet.test.id
-
-      load_balancer_backend_address_pool_ids = [azurerm_lb_backend_address_pool.test.id]
-    }
-  }
-
-  os_disk {
-    storage_account_type = "Standard_LRS"
-    caching              = "ReadWrite"
-  }
-
-  source_image_reference {
-    publisher = "Canonical"
-    offer     = "UbuntuServer"
-    sku       = "16.04-LTS"
-    version   = "latest"
-  }
-}
-`, data.RandomInteger, data.Locations.Primary)
-}
-
-func (OrchestratedVirtualMachineScaleSetResource) priorityTemplate(data acceptance.TestData) string {
-	r := OrchestratedVirtualMachineScaleSetResource{}
-	return fmt.Sprintf(`
-provider "azurerm" {
-  features {}
-}
-
-resource "azurerm_resource_group" "test" {
-  name     = "acctestRG-OVMSS-%[1]d"
-  location = "%[2]s"
-}
-
-%[3]s
-
-resource "azurerm_orchestrated_virtual_machine_scale_set" "test" {
-  name                = "acctestOVMSS-%[1]d"
-  location            = azurerm_resource_group.test.location
-  resource_group_name = azurerm_resource_group.test.name
-
-  priority        = "Spot"
-  eviction_policy = "Deallocate"
-
-  sku_name  = "Standard_D1_v2"
-  instances = 1
-
-  platform_fault_domain_count = 2
-
-  os_profile {
-    linux_configuration {
-      computer_name_prefix = "testvm-%[1]d"
-      admin_username       = "myadmin"
-      admin_password       = "Passwword1234"
-
-      disable_password_authentication = false
-    }
-  }
-
-  network_interface {
-    name    = "TestNetworkProfile"
-    primary = true
-
-    ip_configuration {
-      name      = "TestIPConfiguration"
-      primary   = true
-      subnet_id = azurerm_subnet.test.id
-
-      public_ip_address {
-        name                    = "TestPublicIPConfiguration"
-        domain_name_label       = "test-domain-label"
-        idle_timeout_in_minutes = 4
-      }
-    }
-  }
-
-  os_disk {
-    storage_account_type = "Standard_LRS"
-    caching              = "ReadWrite"
-  }
-
-  source_image_reference {
-    publisher = "Canonical"
-    offer     = "UbuntuServer"
-    sku       = "16.04-LTS"
-    version   = "latest"
-  }
-}
-`, data.RandomInteger, data.Locations.Primary, r.natgateway_template(data))
-}
-
-func (OrchestratedVirtualMachineScaleSetResource) extensionTemplate(data acceptance.TestData) string {
-	r := OrchestratedVirtualMachineScaleSetResource{}
-	return fmt.Sprintf(`
-provider "azurerm" {
-  features {}
-}
-
-resource "azurerm_resource_group" "test" {
-  name     = "acctestRG-OVMSS-%[1]d"
-  location = "%[2]s"
-}
-
-%[3]s
-
-resource "azurerm_orchestrated_virtual_machine_scale_set" "test" {
-  name                = "acctestOVMSS-%[1]d"
-  location            = azurerm_resource_group.test.location
-  resource_group_name = azurerm_resource_group.test.name
-
-  sku_name  = "Standard_D1_v2"
-  instances = 1
-
-  platform_fault_domain_count = 2
-
-  os_profile {
-    linux_configuration {
-      computer_name_prefix = "testvm-%[1]d"
-      admin_username       = "myadmin"
-
-      admin_ssh_key {
-        username   = "myadmin"
-        public_key = "ssh-rsa AAAAB3NzaC1yc2EAAAADAQABAAACAQDCsTcryUl51Q2VSEHqDRNmceUFo55ZtcIwxl2QITbN1RREti5ml/VTytC0yeBOvnZA4x4CFpdw/lCDPk0yrH9Ei5vVkXmOrExdTlT3qI7YaAzj1tUVlBd4S6LX1F7y6VLActvdHuDDuXZXzCDd/97420jrDfWZqJMlUK/EmCE5ParCeHIRIvmBxcEnGfFIsw8xQZl0HphxWOtJil8qsUWSdMyCiJYYQpMoMliO99X40AUc4/AlsyPyT5ddbKk08YrZ+rKDVHF7o29rh4vi5MmHkVgVQHKiKybWlHq+b71gIAUQk9wrJxD+dqt4igrmDSpIjfjwnd+l5UIn5fJSO5DYV4YT/4hwK7OKmuo7OFHD0WyY5YnkYEMtFgzemnRBdE8ulcT60DQpVgRMXFWHvhyCWy0L6sgj1QWDZlLpvsIvNfHsyhKFMG1frLnMt/nP0+YCcfg+v1JYeCKjeoJxB8DWcRBsjzItY0CGmzP8UYZiYKl/2u+2TgFS5r7NWH11bxoUzjKdaa1NLw+ieA8GlBFfCbfWe6YVB9ggUte4VtYFMZGxOjS2bAiYtfgTKFJv+XqORAwExG6+G2eDxIDyo80/OA9IG7Xv/jwQr7D6KDjDuULFcN/iTxuttoKrHeYz1hf5ZQlBdllwJHYx6fK2g8kha6r2JIQKocvsAXiiONqSfw== hello@world.com"
-      }
-    }
-  }
-
-  network_interface {
-    name    = "TestNetworkProfile"
-    primary = true
-
-    ip_configuration {
-      name      = "TestIPConfiguration"
-      primary   = true
-      subnet_id = azurerm_subnet.test.id
-
-      public_ip_address {
-        name                    = "TestPublicIPConfiguration"
-        domain_name_label       = "test-domain-label"
-        idle_timeout_in_minutes = 4
-      }
-    }
-  }
-
-  os_disk {
-    storage_account_type = "Standard_LRS"
-    caching              = "ReadWrite"
-  }
-
-  source_image_reference {
-    publisher = "Canonical"
-    offer     = "UbuntuServer"
-    sku       = "16.04-LTS"
-    version   = "latest"
-  }
-
-  extension {
-    name                       = "CustomScript"
-    publisher                  = "Microsoft.Azure.Extensions"
-    type                       = "CustomScript"
-    type_handler_version       = "2.0"
-    auto_upgrade_minor_version = true
-
-    settings = jsonencode({
-      "commandToExecute" = "echo $HOSTNAME"
-    })
-
-    protected_settings = jsonencode({
-      "managedIdentity" = {}
-    })
-
-  }
-}
-`, data.RandomInteger, data.Locations.Primary, r.natgateway_template(data))
-}
-
-func (OrchestratedVirtualMachineScaleSetResource) extensionTemplateUpdated(data acceptance.TestData) string {
-	r := OrchestratedVirtualMachineScaleSetResource{}
-	return fmt.Sprintf(`
-provider "azurerm" {
-  features {}
-}
-
-resource "azurerm_resource_group" "test" {
-  name     = "acctestRG-OVMSS-%[1]d"
-  location = "%[2]s"
-}
-
-%[3]s
-
-resource "azurerm_orchestrated_virtual_machine_scale_set" "test" {
-  name                = "acctestOVMSS-%[1]d"
-  location            = azurerm_resource_group.test.location
-  resource_group_name = azurerm_resource_group.test.name
-
-  sku_name  = "Standard_D1_v2"
-  instances = 1
-
-  platform_fault_domain_count = 2
-
-  os_profile {
-    linux_configuration {
-      computer_name_prefix = "testvm-%[1]d"
-      admin_username       = "myadmin"
-
-      admin_ssh_key {
-        username   = "myadmin"
-        public_key = "ssh-rsa AAAAB3NzaC1yc2EAAAADAQABAAACAQDCsTcryUl51Q2VSEHqDRNmceUFo55ZtcIwxl2QITbN1RREti5ml/VTytC0yeBOvnZA4x4CFpdw/lCDPk0yrH9Ei5vVkXmOrExdTlT3qI7YaAzj1tUVlBd4S6LX1F7y6VLActvdHuDDuXZXzCDd/97420jrDfWZqJMlUK/EmCE5ParCeHIRIvmBxcEnGfFIsw8xQZl0HphxWOtJil8qsUWSdMyCiJYYQpMoMliO99X40AUc4/AlsyPyT5ddbKk08YrZ+rKDVHF7o29rh4vi5MmHkVgVQHKiKybWlHq+b71gIAUQk9wrJxD+dqt4igrmDSpIjfjwnd+l5UIn5fJSO5DYV4YT/4hwK7OKmuo7OFHD0WyY5YnkYEMtFgzemnRBdE8ulcT60DQpVgRMXFWHvhyCWy0L6sgj1QWDZlLpvsIvNfHsyhKFMG1frLnMt/nP0+YCcfg+v1JYeCKjeoJxB8DWcRBsjzItY0CGmzP8UYZiYKl/2u+2TgFS5r7NWH11bxoUzjKdaa1NLw+ieA8GlBFfCbfWe6YVB9ggUte4VtYFMZGxOjS2bAiYtfgTKFJv+XqORAwExG6+G2eDxIDyo80/OA9IG7Xv/jwQr7D6KDjDuULFcN/iTxuttoKrHeYz1hf5ZQlBdllwJHYx6fK2g8kha6r2JIQKocvsAXiiONqSfw== hello@world.com"
-      }
-    }
-  }
-
-  network_interface {
-    name    = "TestNetworkProfile"
-    primary = true
-
-    ip_configuration {
-      name      = "TestIPConfiguration"
-      primary   = true
-      subnet_id = azurerm_subnet.test.id
-
-      public_ip_address {
-        name                    = "TestPublicIPConfiguration"
-        domain_name_label       = "test-domain-label"
-        idle_timeout_in_minutes = 4
-      }
-    }
-  }
-
-  os_disk {
-    storage_account_type = "Standard_LRS"
-    caching              = "ReadWrite"
-  }
-
-  source_image_reference {
-    publisher = "Canonical"
-    offer     = "UbuntuServer"
-    sku       = "16.04-LTS"
-    version   = "latest"
-  }
-
-  extension {
-    name                       = "CustomScript"
-    publisher                  = "Microsoft.Azure.Extensions"
-    type                       = "CustomScript"
-    type_handler_version       = "2.0"
-    auto_upgrade_minor_version = true
-
-    settings = jsonencode({
-      "commandToExecute" = "echo $HOSTNAME"
-      "timestamp"        = "1234567890"
-    })
-
-    protected_settings = jsonencode({
-      "managedIdentity" = {}
-    })
-  }
-}
-`, data.RandomInteger, data.Locations.Primary, r.natgateway_template(data))
-}
-
-func (OrchestratedVirtualMachineScaleSetResource) multipleExtensionsTemplate(data acceptance.TestData) string {
-	r := OrchestratedVirtualMachineScaleSetResource{}
-	return fmt.Sprintf(`
-provider "azurerm" {
-  features {}
-}
-
-resource "azurerm_resource_group" "test" {
-  name     = "acctestRG-OVMSS-%[1]d"
-  location = "%[2]s"
-}
-
-%[3]s
-
-resource "azurerm_orchestrated_virtual_machine_scale_set" "test" {
-  name                = "acctestOVMSS-%[1]d"
-  location            = azurerm_resource_group.test.location
-  resource_group_name = azurerm_resource_group.test.name
-
-  sku_name  = "Standard_D1_v2"
-  instances = 1
-
-  platform_fault_domain_count = 2
-
-  os_profile {
-    linux_configuration {
-      computer_name_prefix = "testvm-%[1]d"
-      admin_username       = "myadmin"
-      admin_password       = "Passwword1234"
-
-      disable_password_authentication = false
-    }
-  }
-
-  network_interface {
-    name    = "TestNetworkProfile"
-    primary = true
-
-    ip_configuration {
-      name      = "TestIPConfiguration"
-      primary   = true
-      subnet_id = azurerm_subnet.test.id
-
-      public_ip_address {
-        name                    = "TestPublicIPConfiguration"
-        domain_name_label       = "test-domain-label"
-        idle_timeout_in_minutes = 4
-      }
-    }
-  }
-
-  os_disk {
-    storage_account_type = "Standard_LRS"
-    caching              = "ReadWrite"
-  }
-
-  source_image_reference {
-    publisher = "Canonical"
-    offer     = "UbuntuServer"
-    sku       = "16.04-LTS"
-    version   = "latest"
-  }
-
-  extension {
-    name                       = "CustomScript"
-    publisher                  = "Microsoft.Azure.Extensions"
-    type                       = "CustomScript"
-    type_handler_version       = "2.0"
-    auto_upgrade_minor_version = true
-
-    settings = jsonencode({
-      "commandToExecute" = "echo $HOSTNAME"
-    })
-
-    protected_settings = jsonencode({
-      "managedIdentity" = {}
-    })
-  }
-
-  extension {
-    name                       = "Docker"
-    publisher                  = "Microsoft.Azure.Extensions"
-    type                       = "DockerExtension"
-    type_handler_version       = "1.0"
-    auto_upgrade_minor_version = true
-  }
-}
-`, data.RandomInteger, data.Locations.Primary, r.natgateway_template(data))
-}
-
-func (OrchestratedVirtualMachineScaleSetResource) multipleExtensionsTemplate_provisionMultipleExtensionOnExistingOvmss(data acceptance.TestData) string {
-	r := OrchestratedVirtualMachineScaleSetResource{}
-	return fmt.Sprintf(`
-provider "azurerm" {
-  features {}
-}
-
-resource "azurerm_resource_group" "test" {
-  name     = "acctestRG-OVMSS-%[1]d"
-  location = "%[2]s"
-}
-
-%[3]s
-
-resource "azurerm_orchestrated_virtual_machine_scale_set" "test" {
-  name                = "acctestOVMSS-%[1]d"
-  location            = azurerm_resource_group.test.location
-  resource_group_name = azurerm_resource_group.test.name
-
-  sku_name  = "Standard_D1_v2"
-  instances = 2
-
-  platform_fault_domain_count = 2
-
-  os_profile {
-    linux_configuration {
-      computer_name_prefix = "testvm-%[1]d"
-      admin_username       = "myadmin"
-      admin_password       = "Passwword1234"
-
-      disable_password_authentication = false
-    }
-  }
-
-  network_interface {
-    name    = "TestNetworkProfile"
-    primary = true
-
-    ip_configuration {
-      name      = "TestIPConfiguration"
-      primary   = true
-      subnet_id = azurerm_subnet.test.id
-
-      public_ip_address {
-        name                    = "TestPublicIPConfiguration"
-        domain_name_label       = "test-domain-label"
-        idle_timeout_in_minutes = 4
-      }
-    }
-  }
-
-  os_disk {
-    storage_account_type = "Standard_LRS"
-    caching              = "ReadWrite"
-  }
-
-  source_image_reference {
-    publisher = "Canonical"
-    offer     = "UbuntuServer"
-    sku       = "16.04-LTS"
-    version   = "latest"
-  }
-
-  extension {
-    name                       = "CustomScript"
-    publisher                  = "Microsoft.Azure.Extensions"
-    type                       = "CustomScript"
-    type_handler_version       = "2.0"
-    auto_upgrade_minor_version = true
-
-    settings = jsonencode({
-      "commandToExecute" = "echo $HOSTNAME"
-    })
-
-    protected_settings = jsonencode({
-      "managedIdentity" = {}
-    })
-  }
-
-  extension {
-    name                       = "Docker"
-    publisher                  = "Microsoft.Azure.Extensions"
-    type                       = "DockerExtension"
-    type_handler_version       = "1.0"
-    auto_upgrade_minor_version = true
-    provision_after_extensions = ["CustomScript"]
-  }
-}
-`, data.RandomInteger, data.Locations.Primary, r.natgateway_template(data))
-}
-
-func (OrchestratedVirtualMachineScaleSetResource) loadBalancerTemplateManagedDataDisks(data acceptance.TestData) string {
-	return fmt.Sprintf(`
-provider "azurerm" {
-  features {}
-}
-
-resource "azurerm_resource_group" "test" {
-  name     = "acctestRG-OVMSS-%[1]d"
-  location = "%[2]s"
-}
-
-resource "azurerm_virtual_network" "test" {
-  name                = "acctvn-%[1]d"
-  address_space       = ["10.0.0.0/16"]
-  location            = azurerm_resource_group.test.location
-  resource_group_name = azurerm_resource_group.test.name
-}
-
-resource "azurerm_subnet" "test" {
-  name                 = "acctsub-%[1]d"
-  resource_group_name  = azurerm_resource_group.test.name
-  virtual_network_name = azurerm_virtual_network.test.name
-  address_prefixes     = ["10.0.2.0/24"]
-}
-
-resource "azurerm_lb" "test" {
-  name                = "acctestlb-%[1]d"
-  location            = azurerm_resource_group.test.location
-  resource_group_name = azurerm_resource_group.test.name
-
-  sku = "Standard"
-
-  frontend_ip_configuration {
-    name                          = "default"
-    subnet_id                     = azurerm_subnet.test.id
-    private_ip_address_allocation = "Dynamic"
-  }
-}
-
-resource "azurerm_lb_backend_address_pool" "test" {
-  name            = "test"
-  loadbalancer_id = azurerm_lb.test.id
-}
-
-resource "azurerm_orchestrated_virtual_machine_scale_set" "test" {
-  name                = "acctestOVMSS-%[1]d"
-  location            = azurerm_resource_group.test.location
-  resource_group_name = azurerm_resource_group.test.name
-
-  sku_name  = "Standard_F2"
-  instances = 1
-
-  platform_fault_domain_count = 2
-
-  os_profile {
-    linux_configuration {
-      computer_name_prefix = "testvm-%[1]d"
-      admin_username       = "myadmin"
-      admin_password       = "Passwword1234"
-
-      disable_password_authentication = false
-    }
-  }
-
-  network_interface {
-    name    = "TestNetworkProfile"
-    primary = true
-
-    ip_configuration {
-      name      = "TestIPConfiguration"
-      primary   = true
-      subnet_id = azurerm_subnet.test.id
-
-      load_balancer_backend_address_pool_ids = [azurerm_lb_backend_address_pool.test.id]
-    }
-  }
-
-  os_disk {
-    storage_account_type = "Standard_LRS"
-    caching              = "ReadWrite"
-  }
-
-  data_disk {
-    lun                  = 0
-    caching              = "ReadWrite"
-    create_option        = "Empty"
-    disk_size_gb         = 10
-    storage_account_type = "Standard_LRS"
-  }
-
-  source_image_reference {
-    publisher = "Canonical"
-    offer     = "UbuntuServer"
-    sku       = "16.04.0-LTS"
-    version   = "latest"
-  }
-}
-`, data.RandomInteger, data.Locations.Primary)
-}
-
-func (OrchestratedVirtualMachineScaleSetResource) nonStandardCasing(data acceptance.TestData) string {
-	r := OrchestratedVirtualMachineScaleSetResource{}
-	return fmt.Sprintf(`
-provider "azurerm" {
-  features {}
-}
-
-resource "azurerm_resource_group" "test" {
-  name     = "acctestRG-OVMSS-%[1]d"
-  location = "%[2]s"
-}
-
-%[3]s
-
-resource "azurerm_orchestrated_virtual_machine_scale_set" "test" {
-  name                = "acctestOVMSS-%[1]d"
-  location            = azurerm_resource_group.test.location
-  resource_group_name = azurerm_resource_group.test.name
-
-  sku_name  = "Standard_F2"
-  instances = 2
-
-  platform_fault_domain_count = 2
-
-  os_profile {
-    linux_configuration {
-      computer_name_prefix = "testvm-%[1]d"
-      admin_username       = "myadmin"
-      admin_password       = "Passwword1234"
-
-      disable_password_authentication = false
-    }
-  }
-
-  network_interface {
-    name    = "TestNetworkProfile-%[1]d"
-    primary = true
-
-    ip_configuration {
-      name      = "TestIPConfiguration"
-      primary   = true
-      subnet_id = azurerm_subnet.test.id
-
-      public_ip_address {
-        name                    = "TestPublicIPConfiguration"
-        domain_name_label       = "test-domain-label"
-        idle_timeout_in_minutes = 4
-      }
-    }
-  }
-
-  os_disk {
-    storage_account_type = "Standard_LRS"
-    caching              = "ReadWrite"
-  }
-
-  source_image_reference {
-    publisher = "Canonical"
-    offer     = "UbuntuServer"
-    sku       = "16.04-LTS"
-    version   = "latest"
-  }
-}
-`, data.RandomInteger, data.Locations.Primary, r.natgateway_template(data))
-}
-
-func (OrchestratedVirtualMachineScaleSetResource) multipleNetworkProfiles(data acceptance.TestData) string {
-	r := OrchestratedVirtualMachineScaleSetResource{}
-	return fmt.Sprintf(`
-provider "azurerm" {
-  features {}
-}
-
-resource "azurerm_resource_group" "test" {
-  name     = "acctestRG-OVMSS-%[1]d"
-  location = "%[2]s"
-}
-
-%[3]s
-
-resource "azurerm_orchestrated_virtual_machine_scale_set" "test" {
-  name                = "acctestOVMSS-%[1]d"
-  location            = azurerm_resource_group.test.location
-  resource_group_name = azurerm_resource_group.test.name
-
-  sku_name  = "Standard_D1_v2"
-  instances = 2
-
-  platform_fault_domain_count = 2
-
-  os_profile {
-    linux_configuration {
-      computer_name_prefix = "testvm-%[1]d"
-      admin_username       = "myadmin"
-      admin_password       = "Passwword1234"
-
-      disable_password_authentication = false
-    }
-  }
-
-  network_interface {
-    name    = "primary-%[1]d"
-    primary = true
-
-    ip_configuration {
-      name      = "primary"
-      primary   = true
-      subnet_id = azurerm_subnet.test.id
-
-      public_ip_address {
-        name                    = "TestPublicIPConfiguration"
-        domain_name_label       = "test-domain-label"
-        idle_timeout_in_minutes = 4
-      }
-    }
-  }
-
-  network_interface {
-    name    = "secondary-%[1]d"
-    primary = false
-
-    ip_configuration {
-      name      = "secondary"
-      primary   = true
-      subnet_id = azurerm_subnet.test.id
-    }
-  }
-
-  os_disk {
-    storage_account_type = "Standard_LRS"
-    caching              = "ReadWrite"
-  }
-
-  source_image_reference {
-    publisher = "Canonical"
-    offer     = "UbuntuServer"
-    sku       = "16.04-LTS"
-    version   = "latest"
   }
 }
 `, data.RandomInteger, data.Locations.Primary, r.natgateway_template(data))
